fix(memory): allow verified generation state advancement
